+++ qlat/field-serial-io.h
@@ -226,7 +226,7 @@ crc32_t field_crc32(const Field<M>& f)
   crc32_t ret = 0;
   for (int i = 0; i < std::min(num_node, new_num_node); ++i) {
     if (i == id_node) {
-      for (int k = 0; k < fs.size(); ++k) {
+      for (int k = 0; k < (int)fs.size(); ++k) {
         ret = crc32_par(ret, get_data(fs[k]));
       }
     }
